Created individual component for AddToCart feature, added AddToCart to SingleProduct and Products components. Also showed a message on SingleProduct page if no reviews exist
