+++ client/components/SingleProduct.js
@@ -1,68 +1,24 @@
 import React from 'react'
 import {getProduct} from '../store/product'
 import {connect} from 'react-redux'
-import SingleReview from './SingleReview'
-import {editCart} from '../store/cart'
+import {SingleReview, AddToCart} from './index'
 
 const mapStateToProps = state => {
   return {
-    product: state.product,
-    cart: state.cart
+    product: state.product
   }
 }
 
 const mapDispatchToProps = dispatch => {
   return {
-    getProduct: productId => dispatch(getProduct(productId)),
-    editCart: products => dispatch(editCart(products))
+    getProduct: productId => dispatch(getProduct(productId))
   }
 }
 
 class SingleProduct extends React.Component {
-  constructor() {
-    super()
-    this.state = {
-      quantity: 1
-    }
-    this.increment = this.increment.bind(this)
-    this.decrement = this.decrement.bind(this)
-    this.addToCart = this.addToCart.bind(this)
-  }
   componentDidMount() {
     this.props.getProduct(this.props.match.params.productId)
   }
-
-  addToCart() {
-    const {product, cart, editCart} = this.props
-    const newCart = [...cart]
-    const {quantity} = this.state
-    const prodIndex = cart.findIndex(item => {
-      return item.product.id === product.id
-    })
-    if (prodIndex > -1) {
-      newCart[prodIndex] = {
-        product,
-        quantity: newCart[prodIndex].quantity + quantity
-      }
-      editCart(newCart)
-    } else {
-      editCart([...cart, {product, quantity}])
-    }
-  }
-
-  increment() {
-    this.setState({quantity: this.state.quantity + 1})
-  }
-  decrement() {
-    if (this.state.quantity > 1) {
-      this.setState({quantity: this.state.quantity - 1})
-    }
-  }
-
-  handleChange(event) {
-    this.setState({quantity: event.target.value})
-  }
-
   render() {
     const {product} = this.props
     return (
@@ -74,27 +30,14 @@ class SingleProduct extends React.Component {
           <h2>{product.name}</h2>
           <p>{product.description}</p>
         </div>
-        <div className="add-product">
-          <button type="button" onClick={this.addToCart}>
-            Add to cart
-          </button>
-          <input
-            className="quantity"
-            onChange={this.handleChange}
-            value={this.state.quantity}
-            type="number"
-          />
-          <button type="button" className="plus" onClick={this.increment}>
-            +
-          </button>
-          <button type="button" className="minus" onClick={this.decrement}>
-            -
-          </button>
-        </div>
-        {product.reviews &&
+        <AddToCart product={product} />
+        {product.reviews ? (
           product.reviews.map(review => (
             <SingleReview key={review.id} review={review} />
-          ))}
+          ))
+        ) : (
+          <h2>There are no reviews for this product</h2>
+        )}
       </div>
     )
   }
