+++ client/components/index.js
@@ -10,4 +10,6 @@ export {default as AdminProducts} from './admin-products'
 export {Login, Signup} from './auth-form'
 export {default as Products} from './products'
 export {default as SingleProduct} from './SingleProduct'
+export {default as SingleReview} from './SingleReview'
 export {default as Cart} from './Cart'
+export {default as AddToCart} from './AddToCart'
